♻️ Split $lib/metadata into multiple files
