+++ src/lib/observations.js
@@ -1,4 +1,4 @@
-import { generateId } from '$lib/database.js';
+import { generateId, Schemas } from '$lib/database.js';
 
 import * as db from './idb.svelte.js';
 import { tables } from './idb.svelte.js';
@@ -6,6 +6,11 @@ import { deleteImageFile, imageFileIds } from './images.js';
 import { mergeMetadataValues } from './metadata/index.js';
 import { uiState } from './state.svelte.js';
 import { compareBy, nonnull } from './utils.js';
+
+/**
+ * @import * as DB from '$lib/database.js'
+ * @import { DatabaseHandle } from '$lib/idb.svelte.js'
+ */
 
 /**
  * @param {string[]} parts IDs of observations or images to merge
@@ -177,4 +182,30 @@ export async function ensureNoLoneImages(tx) {
 			}
 		}
 	});
+}
+
+/**
+ * Gets all metadata for an observation, including metadata derived from merging the metadata values of the images that make up the observation.
+ * @param {Pick<DB.Observation, 'images' | 'metadataOverrides'>} observation
+ * @param {DatabaseHandle} db
+ * @param {DB.Protocol} protocol
+ * @returns {Promise<DB.MetadataValues>}
+ */
+export async function observationMetadata(db, protocol, observation) {
+	const images = await Promise.all(
+		observation.images.map(async (id) => await db.get('Image', id))
+	).then((ims) => ims.filter(nonnull));
+
+	images.sort(compareBy(({ id }) => observation.images.indexOf(id)));
+
+	const metadataFromImages = await mergeMetadataValues(
+		db,
+		protocol,
+		images.map((img) => Schemas.MetadataValues.assert(img.metadata))
+	);
+
+	return {
+		...metadataFromImages,
+		...observation.metadataOverrides
+	};
 }
